% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doTmleUpdate.R
\name{applyFinalSubmodel}
\alias{applyFinalSubmodel}
\title{applyFinalSubmodel}
\usage{
applyFinalSubmodel(
  Estimates,
  final_eps,
  Data,
  TargetEvent,
  TargetTime,
  Verbose = FALSE
)
}
\arguments{
\item{Estimates}{the real estimates list to be updated in-place}

\item{final_eps}{numeric vector from root solver}

\item{Data}{data.table}

\item{TargetEvent}{numeric}

\item{TargetTime}{numeric}

\item{Verbose}{boolean}
}
\value{
Updated \code{Estimates} object with final hazards + EIC
}
\description{
Commits the final epsVec solution from nleqslv onto the real 'Estimates'.
}
