% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConCR-TMLE.R
\docType{package}
\name{ConCR-TMLE}
\alias{ConCR-TMLE}
\title{ConCR-TMLE}
\description{
Implements continuous-time TMLE for point binary treatment right censored survival (with competing risks)
}
\section{ConCR-TMLE major sub-functions}{

formatArguments()
many checkXXXXXX() functions
getInitialEstimates()
getPropScores()
getHazEstimates()
getEIC()
getIC()
doTMLEUpdate()
}

