% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contmle.R
\name{contmle}
\alias{contmle}
\title{contmle}
\usage{
contmle(
  dt,
  estimation = list(outcome = list(fit = c("sl", "sl", "hal", "km"), model = Surv(time,
    delta == 1) ~ A + L1 + L2 + L3, changepoint = NULL), cens = list(fit = c("sl", "sl",
    "hal", "km"), model = Surv(time, delta == 0) ~ L2 + L3 + A * L1, changepoint = NULL)),
  target = 1,
  hal.screening = FALSE,
  one.step = FALSE,
  deps.size = 0.1,
  no.small.steps = 200,
  push.criterion = FALSE,
  iterative = FALSE,
  no.update.if.solved = FALSE,
  separate.cr = FALSE,
  simultaneous.ci = FALSE,
  treat.model = A ~ L1 + L2 + L3,
  cut.off.cens = 0.001,
  weighted.norm = c(FALSE, "sigma", "Sigma"),
  treat.effect = c("1", "0", "ate", "stochastic"),
  pi.star.fun = function(L) L$L1 * 0.2,
  tau = c(1.2),
  use.observed.times = FALSE,
  length.times = length(tau),
  check.times.size = NULL,
  delta.min = 0.01,
  check.min = min(tau) + 0.1,
  check.max = max(tau) - 0.1,
  sl.method = 3,
  V = 5,
  lambda.cv = NULL,
  maxit = 1000,
  init.lambda.cvs = c(sapply(1:5, function(jjj) (9:1)/(10^jjj))),
  lambda.cvs = seq(1e-07, 0.01, length = 50),
  lambda.grid.size = 50,
  penalize.time = FALSE,
  cut.covars = 8,
  cut.time = 10,
  cut.time.A = 10,
  cut.L.A = 8,
  cut.L.interaction = 3,
  maxIter = 10,
  verbose = FALSE,
  verbose.sl = FALSE,
  check.sup = FALSE,
  output.km = FALSE,
  only.km = FALSE,
  only.cox.sl = FALSE,
  output.RR = FALSE,
  sl.change.points = (0:12)/10,
  sl.models = list(mod1 = list(Surv(time, delta == 1) ~ A + L1 + L2 + L3), mod2 =
    list(Surv(time, delta == 1) ~ A + L1.squared + L2 + L3), mod3 = list(Surv(time, delta
    == 1) ~ L2.squared + A + L1.squared + L2 + L3), mod4 = list(Surv(time, delta == 1) ~
    A + L1.squared), mod5 = list(Surv(time, delta == 1) ~ A * L1 + L2 + L3), mod6 =
    list(Surv(time, delta == 1) ~ A * L1.squared + L2 + L3))
)
}
\arguments{
\item{estimation}{outcome model}

\item{target}{when there are competing risks, what is the target? if 0, only survival is targeted}

\item{hal.screening}{use poisson-based or cox-based hal?}

\item{one.step}{use iterative or one-step tmle; (for competing risks, one-step is default)}

\item{no.update.if.solved}{option to not update in directions currently already solved}

\item{separate.cr}{target cause-specific hazards separately or together in one-step tmle}

\item{simultaneous.ci}{simultaneous inference}

\item{treat.model}{treatment model}

\item{cut.off.cens}{cut-off for truncation of censoring weights}

\item{weighted.norm}{use weighted norm in multivariate one-step}

\item{treat.effect}{treatment effect of interest}

\item{pi.star.fun}{specify stochastic intervention}

\item{tau}{time-point(s) of interest}

\item{use.observed.times}{increasing grid?}

\item{sl.method}{pick super learning loss (should not change this)}

\item{V}{number of folds in cross-validation}

\item{lambda.cv}{specify penalization in hal?}

\item{init.lambda.cvs}{specify grid over which to pick penalization in hal by cross-validation}

\item{penalize.time}{penalize time indicators in hal?}

\item{cut.covars}{pick grid for indicators in hal}

\item{cut.time}{pick grid for indicators in hal}

\item{cut.time.A}{pick grid for indicators in hal}

\item{cut.L.A}{pick grid for indicators in hal}

\item{cut.L.interaction}{pick grid for indicators in hal}

\item{maxIter}{maximum number of iterations in iterative tmle}

\item{output.km}{for comparison; output kaplan-meier and hr}

\item{only.km}{for comparison; output kaplan-meier and hr}

\item{only.cox.sl}{for comparison; output kaplan-meier and hr}

\item{sl.change.points}{models incorporated in super learner}

\item{sl.models}{models incorporated in super learner}
}
\value{

}
\description{
contmle
}
