% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.loss.fun.R
\name{cox.loss.fun}
\alias{cox.loss.fun}
\title{title}
\usage{
cox.loss.fun(
  train.fit,
  dt,
  risk.set,
  test.set,
  X = NULL,
  lambda.cv = NULL,
  delta.var = "delta",
  delta.value = 1,
  change.point = NULL
)
}
\arguments{
\item{train.fit}{model fitted to training data.}

\item{dt}{dataset.}

\item{risk.set}{risk.set used for partial likelihood.}

\item{test.set}{validation data.}

\item{X}{design matrix if one of c("coxnet", "cv.glmnet", "glmnet") is used.}

\item{lambda.cv}{grid over which to choose penalization if one of c("coxnet", "cv.glmnet", "glmnet") is used.}

\item{delta.var}{name of event type variable.}

\item{delta.value}{type of event of interest here-}

\item{change.point}{specified if there is a changepoint in the effect of treatment across time.}
}
\value{
a
}
\description{
title
}
\examples{
c
}
\seealso{
b
}
\author{
Helene C. W. Rytgaard \href{mailto:hely@biostat.ku.dk}{hely@biostat.ku.dk}
}
