% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.sl.R
\name{cox.sl}
\alias{cox.sl}
\title{atitle}
\usage{
cox.sl(
  loss.fun,
  dt,
  V = 5,
  seed = 19192,
  delta.var = NULL,
  delta.value = NULL,
  treatment = "A",
  cox.models = list(mod1 = list(Surv(time, delta == 1) ~ A + L1 + L2 + L3), mod2 =
    list(Surv(time, delta == 1) ~ A + L1.squared + L2 + L3), mod3 = list(Surv(time, delta
    == 1) ~ L2.squared + A + L1.squared + L2 + L3), mod4 = list(Surv(time, delta == 1) ~
    A + L1.squared), mod5 = list(Surv(time, delta == 1) ~ A * L1 + L2 + L3), mod6 =
    list(Surv(time, delta == 1) ~ A * L1.squared + L2 + L3))
)
}
\arguments{
\item{loss.fun}{should just be set to cox.loss.fun.}

\item{dt}{dataset.}

\item{V}{number of folds in cross-validation.}

\item{seed}{random seed :).}

\item{delta.var}{name of event type variable.}

\item{delta.value}{type of event of interest here-}

\item{treatment}{name of treatment variable.}

\item{cox.models}{a list of Cox models to be compared with cross-validation.}

\item{change.points}{specified if there is a changepoint in the effect of treatment across time.}
}
\value{
a
}
\description{
bla
}
\details{
blub
}
\examples{
c
}
\seealso{
b
}
\author{
Helene C. W. Rytgaard \href{mailto:hely@biostat.ku.dk}{hely@biostat.ku.dk}
}
