% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.fun.R
\name{cv.fun}
\alias{cv.fun}
\title{atitle}
\usage{
cv.fun(
  loss.fun,
  dt,
  V = 5,
  seed = 19192,
  X = NULL,
  Y = NULL,
  offset = NULL,
  method.risk = c("test", "train", "VvH"),
  time.var = NULL,
  penalty.factor = rep(1, ncol(X)),
  delta.var = "delta",
  delta.value = 1,
  change.point = NULL,
  treatment = NULL,
  cox.model = NULL,
  lambda.cvs = c(sapply(1:5, function(jjj) (9:1)/(10^jjj)))
)
}
\arguments{
\item{loss.fun}{should be set to cox.loss.fun if comparing cox models;}

\item{dt}{dataset.}

\item{V}{number of folds in cross-validation.}

\item{seed}{random seed :).}

\item{X}{design matrix if one of c("coxnet", "cv.glmnet", "glmnet") is used.}

\item{Y}{outcome object (Surv)  if one of c("coxnet", "cv.glmnet", "glmnet") is used.}

\item{offset}{offset used to construct poisson hal.}

\item{method.risk}{option to pick different cross-validation schemes for cox models; basically it picks the
risk set for the partial likelihood. Should be chosen as 'test'.}

\item{time.var}{name of time variable.}

\item{penalty.factor}{variable to specify if certain (groups of) variables should not be penalized.}

\item{delta.var}{name of event type variable.}

\item{delta.value}{type of event of interest here-}

\item{change.point}{specified if there is a changepoint in the effect of treatment across time.}

\item{treatment}{name of treatment variable.}

\item{cox.model}{model to compute the cve for.}

\item{lambda.cv}{grid over which to choose penalization if one of c("coxnet", "cv.glmnet", "glmnet") is used.}
}
\value{
a
}
\description{
bla
}
\details{
blub
}
\examples{
c
}
\seealso{
b
}
\author{
Helene C. W. Rytgaard \href{mailto:hely@biostat.ku.dk}{hely@biostat.ku.dk}
}
