% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doTmleUpdate.R
\name{deepcopyEstimates}
\alias{deepcopyEstimates}
\title{Deep-copy an Estimates list so we can do parallel updates safely (Jacobi).}
\usage{
deepcopyEstimates(E)
}
\arguments{
\item{E}{The original Estimates object (list of arms).}
}
\value{
A new, independent copy of E.
}
\description{
Deep-copy an Estimates list so we can do parallel updates safely (Jacobi).
}
\keyword{internal}
