% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInitialEstimate.R
\name{getInitialEstimate}
\alias{getInitialEstimate}
\title{getInitialEstimate}
\usage{
getInitialEstimate(
  Data,
  Model,
  CVFolds,
  MinNuisance,
  TargetEvent,
  TargetTime,
  Regime,
  ReturnModels,
  HazEnsemble = FALSE,
  CensoringTV = NULL
)
}
\arguments{
\item{Data}{data.table}

\item{Model}{list}

\item{CVFolds}{: list}

\item{MinNuisance}{numeric}

\item{TargetEvent}{numeric vector}

\item{TargetTime}{numeric vector}

\item{Regime}{list}

\item{ReturnModels}{boolean}
}
\description{
getInitialEstimate
}
\keyword{internal}
