% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPropScore.R
\name{getPropScore}
\alias{getPropScore}
\title{getPropScore}
\usage{
getPropScore(
  TrtVal,
  CovDT,
  TrtModel,
  MinNuisance,
  Regime,
  CVFolds,
  TrtLoss = NULL,
  ReturnModels,
  PredTrtVal = NULL,
  PredCovDT = NULL,
  PredRegime = NULL
)
}
\arguments{
\item{TrtVal}{numeric vector}

\item{CovDT}{data.table}

\item{TrtModel}{list or fitted object}

\item{MinNuisance}{numeric}

\item{Regime}{list}

\item{CVFolds}{list}

\item{TrtLoss}{character or function(A, g.A)}

\item{ReturnModels}{boolean}
}
\description{
getPropScore
}
\keyword{internal}
