% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPropScores.R
\name{getPropScores}
\alias{getPropScores}
\title{Title}
\usage{
getPropScores(
  Data,
  CovDataTable,
  Models,
  MinNuisance,
  RegsOfInterest,
  PropScoreBackend,
  CVFolds
)
}
\arguments{
\item{Data}{data.table}

\item{CovDataTable}{data.table}

\item{Models}{list}

\item{MinNuisance}{numeric}

\item{RegsOfInterest}{list}

\item{PropScoreBackend}{character}

\item{CVFolds}{list}
}
\description{
Title
}
