% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHazEstimate.R
\name{isSurvivalSLSpec}
\alias{isSurvivalSLSpec}
\title{Check if a model specification is for survivalSL}
\usage{
isSurvivalSLSpec(ModelSpec)
}
\arguments{
\item{ModelSpec}{model specification object}
}
\value{
logical
}
\description{
Check if a model specification is for survivalSL
}
\keyword{internal}
