% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHazEstimate.R
\name{predictSurvivalSLHazard}
\alias{predictSurvivalSLHazard}
\title{Convert survivalSL survival predictions to hazard estimates}
\usage{
predictSurvivalSLHazard(SLFit, newdata, times, CovCols)
}
\arguments{
\item{SLFit}{fitted survivalSL object}

\item{newdata}{data.frame for prediction}

\item{times}{numeric vector of times for hazard estimation}

\item{CovCols}{character vector of covariate column names}
}
\value{
matrix of hazard estimates (times x subjects)
}
\description{
Convert survivalSL survival predictions to hazard estimates
}
\keyword{internal}
