% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doTmleUpdate.R
\name{printOneStepDiagnostics}
\alias{printOneStepDiagnostics}
\title{Print diagnostic information about TMLE convergence}
\usage{
printOneStepDiagnostics(OneStepStop, NormPnEIC)
}
\arguments{
\item{OneStepStop}{data.table with convergence check results}

\item{NormPnEIC}{numeric, the current norm of PnEIC}
}
\description{
Print diagnostic information about TMLE convergence
}
\keyword{internal}
