% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetRMST.R
\name{rmstTargetArm}
\alias{rmstTargetArm}
\title{Fluctuate one arm's hazards to solve the RMST/LYL estimating equations}
\usage{
rmstTargetArm(
  est,
  TargetEvent,
  Horizon,
  T.tilde,
  Delta,
  EvalTimes,
  MaxUpdateIter,
  OneStepEps,
  EICStopRule = "hybrid",
  EICStopAbsTol = 0,
  Verbose,
  ArmName = ""
)
}
\description{
Fluctuate one arm's hazards to solve the RMST/LYL estimating equations
}
\keyword{internal}
