% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimTrueRisks.R
\docType{data}
\name{SimTrueRisks}
\alias{SimTrueRisks}
\title{LEADER covariates with permuted outcomes}
\format{
A data frame with 9340 rows and 15 variables:
\describe{
\item{F.a0}{control risk}
\item{F.a1}{treated risk}
\item{J}{event type}
\item{RD}{risk diff}
\item{RR}{risk ratio}
\item{SR}{surv ratio}
\item{time}{time}
...
}
}
\usage{
SimTrueRisks
}
\description{
LEADER covariates with permuted outcomes
}
\keyword{datasets}
