% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doTmleUpdate.R
\name{updateHazardsWithEps}
\alias{updateHazardsWithEps}
\title{updateHazardsWithEps}
\usage{
updateHazardsWithEps(
  localEst,
  epsVec,
  Data,
  TargetEvent,
  TargetTime,
  Verbose = FALSE
)
}
\arguments{
\item{localEst}{ephemeral copy of the Estimates list}

\item{epsVec}{numeric vector (length = k)}

\item{Data}{data.table}

\item{TargetEvent}{numeric}

\item{TargetTime}{numeric}

\item{Verbose}{boolean}
}
\value{
The updated localEst (ephemeral)
}
\description{
Illustrative helper function that modifies local hazards based on epsVec.
You must define how each eps entry maps to a hazard shift or scale.
}
