% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doConcrete.R
\name{doConcrete}
\alias{doConcrete}
\alias{print.ConcreteEst}
\alias{plot.ConcreteEst}
\title{doConcrete}
\usage{
doConcrete(ConcreteArgs)

\method{print}{ConcreteEst}(x, ...)

\method{plot}{ConcreteEst}(x, convergence = TRUE, propscores = TRUE, ask = TRUE, ...)
}
\arguments{
\item{ConcreteArgs}{"ConcreteArgs" sl3 object : output of formatArguments()}

\item{x}{a ConcreteEst object}

\item{...}{additional arguments to be passed into plot methods}

\item{convergence}{logical: plot the PnEIC norms for each TMLE small update step}

\item{propscores}{logical: plot the densities of the propensity scores for each intervention}

\item{ask}{logical: whether or not to prompt for user input before displaying plots}
}
\value{
object with s3 class "ConcreteEst"
}
\description{
doConcrete
}
\section{Functions}{
\itemize{
\item \code{print(ConcreteEst)}: print.ConcreteEst print method for "ConcreteEst" class

\item \code{plot(ConcreteEst)}: plot.ConcreteEst plot method for "ConcreteEst" class

}}
\examples{
library(data.table)
library(survival)
data <- as.data.table(survival::pbc)
data[, trt := sample(0:1, nrow(data), TRUE)]
cols <- c("id", "time", "status", "trt",
          "age", "albumin", "sex", "bili")
data <- data[, .SD, .SDcols = cols]

intervention <- makeITT()
target.time <- 2500
target.event <- 1:2
model <- list("trt" = c("SL.glm", "SL.glmnet"),
              "0" = list(Surv(time, status == 0) ~ .),
              "1" = list(Surv(time, status == 1) ~ .),
              "2" = list(Surv(time, status == 2) ~ .))

# formatArguments() returns correctly formatted arguments for doConcrete()
concrete.args <- formatArguments(Data = data,
                                 EventTime = "time",
                                 EventType = "status",
                                 Treatment = "trt",
                                 ID = "id",
                                 Intervention = intervention,
                                 TargetTime = target.time,
                                 TargetEvent = target.event,
                                 Model = model,
                                 Verbose = FALSE)

# doConcrete() returns tmle (and g-comp plug-in) estimates of targeted risks
concrete.est <- doConcrete(concrete.args)

# getOutput returns risk difference, relative risk, and treatment-specific risks
concrete.out <- getOutput(Estimate = concrete.est, Estimand = c("rd", "rr", "risk"))
concrete.out$RD
concrete.out$RR
concrete.out$Risk
}
