% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doConCRTmle.R
\name{doConcrete}
\alias{doConcrete}
\title{doConcrete}
\usage{
doConcrete(ConcreteArgs)
}
\arguments{
\item{ConcreteArgs}{: output from concrete::formatArguments}
}
\value{
tbd
}
\description{
doConcrete
}
\examples{
"tbd"
}
