% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doConCRTmle.R
\name{doConCRTmle}
\alias{doConCRTmle}
\title{Title}
\usage{
doConCRTmle(
  EventTime,
  EventType,
  Treatment,
  CovDataTable,
  ID = NULL,
  TargetTimes = sort(unique(EventTime)),
  TargetEvents = NULL,
  Models,
  CVArgs = NULL,
  NumUpdateSteps = 25,
  OneStepEps = 0.1,
  PropScoreCutoff = 0.05,
  Verbose = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
