% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doTmleUpdateFuns.R
\name{doTmleUpdate}
\alias{doTmleUpdate}
\title{Title}
\usage{
doTmleUpdate(
  Estimates,
  SummEIC,
  Data,
  Censored,
  TargetEvents,
  TargetTimes,
  Events,
  NumUpdateSteps,
  OneStepEps,
  NormPnEIC,
  Verbose
)
}
\arguments{
\item{Estimates}{list}

\item{SummEIC}{data.table}

\item{Data}{data.table}

\item{Censored}{boolean}

\item{TargetEvents}{numeric vector}

\item{TargetTimes}{numeric vector}

\item{Events}{numeric vector}

\item{NumUpdateSteps}{numeric}

\item{OneStepEps}{numeric}

\item{NormPnEIC}{numeric}

\item{Verbose}{boolean}
}
\description{
Title
}
