% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doTmleUpdate.R
\name{doTmleUpdate}
\alias{doTmleUpdate}
\title{Title}
\usage{
doTmleUpdate(
  Estimates,
  SummEIC,
  Data,
  TargetEvent,
  TargetTime,
  MaxUpdateIter,
  OneStepEps,
  NormPnEIC,
  Verbose
)
}
\arguments{
\item{Estimates}{list}

\item{SummEIC}{data.table}

\item{Data}{data.table}

\item{TargetEvent}{numeric vector}

\item{TargetTime}{numeric vector}

\item{MaxUpdateIter}{numeric}

\item{OneStepEps}{numeric}

\item{NormPnEIC}{numeric}

\item{Verbose}{boolean}
}
\description{
Title
}
