% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatArguments.R
\name{formatArguments}
\alias{formatArguments}
\title{Title}
\usage{
formatArguments(
  EventTime,
  EventType,
  Treatment,
  Intervention,
  CovDataTable,
  ID,
  TargetTimes,
  TargetEvents,
  Models,
  CVArgs,
  NumUpdateSteps,
  OneStepEps,
  MinNuisance,
  PropScoreBackend,
  Verbose,
  GComp
)
}
\arguments{
\item{EventTime}{: Numeric vector (N x 1)}

\item{EventType}{: Numeric (integer) vector (N x 1)}

\item{Treatment}{: Numeric vector (N x 1)}

\item{Intervention}{: list of function (length = A*)}

\item{CovDataTable}{: data.table (N x ?)}

\item{ID}{: vector (N x 1)}

\item{TargetTimes}{: numeric vector (length = K)}

\item{TargetEvents}{: numeric vector \\subset EventType (length = J)}

\item{Models}{: list of functions (length = L)}

\item{CVArgs}{: list}

\item{NumUpdateSteps}{: numeric}

\item{OneStepEps}{: numeric}

\item{MinNuisance}{: numeric}

\item{PropScoreBackend}{: character}

\item{Verbose}{: boolean}

\item{GComp}{: boolean}
}
\value{
tbd
}
\description{
Title
}
