% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEICFuns.R
\name{getEIC}
\alias{getEIC}
\title{get EICs}
\usage{
getEIC(
  Estimates,
  Data,
  RegsOfInterest,
  Censored,
  TargetEvents,
  TargetTimes,
  Events,
  MinNuisance,
  GComp = FALSE
)
}
\arguments{
\item{Estimates}{list}

\item{Data}{data.table}

\item{RegsOfInterest}{list}

\item{Censored}{boolean}

\item{TargetEvents}{numeric vector}

\item{TargetTimes}{numeric vector}

\item{Events}{numeric vector}

\item{MinNuisance}{numeric}

\item{GComp}{boolean}
}
\description{
get EICs
}
