% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEIC.R
\name{getEIC}
\alias{getEIC}
\title{get EICs}
\usage{
getEIC(
  Estimates,
  Data,
  Regime,
  Censored,
  TargetEvent,
  TargetTime,
  Events,
  MinNuisance,
  GComp = FALSE
)
}
\arguments{
\item{Estimates}{list}

\item{Data}{data.table}

\item{Regime}{list}

\item{Censored}{boolean}

\item{TargetEvent}{numeric vector}

\item{TargetTime}{numeric vector}

\item{Events}{numeric vector}

\item{MinNuisance}{numeric}

\item{GComp}{boolean}
}
\description{
get EICs
}
