% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHazEstimate.R
\name{getHazFit}
\alias{getHazFit}
\title{Title}
\usage{
getHazFit(Data, Models, CVFolds, Hazards)
}
\arguments{
\item{Data}{data.table}

\item{Models}{list}

\item{CVFolds}{list}

\item{Hazards}{list}
}
\description{
Title
}
