% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHazEstimate.R
\name{getHazFit}
\alias{getHazFit}
\title{Title}
\usage{
getHazFit(Data, Model, CVFolds, Hazards, ReturnModels)
}
\arguments{
\item{Data}{data.table}

\item{Model}{list}

\item{CVFolds}{list}

\item{Hazards}{list}

\item{ReturnModels}{boolean}
}
\description{
Title
}
