% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInitialEstimate.R
\name{getInitialEstimate}
\alias{getInitialEstimate}
\title{Title}
\usage{
getInitialEstimate(
  Data,
  CovDataTable,
  Models,
  MinNuisance,
  TargetEvents,
  TargetTimes,
  RegsOfInterest,
  PropScoreBackend,
  Censored
)
}
\arguments{
\item{Data}{data.table}

\item{CovDataTable}{data.table}

\item{Models}{list}

\item{MinNuisance}{numeric}

\item{TargetEvents}{numeric vector}

\item{TargetTimes}{numeric vector}

\item{RegsOfInterest}{list}

\item{PropScoreBackend}{character}

\item{Censored}{boolean}
}
\description{
Title
}
