% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInitialEstimate.R
\name{getInitialEstimate}
\alias{getInitialEstimate}
\title{Title}
\usage{
getInitialEstimate(
  Data,
  CovDataTable,
  Model,
  CVFolds,
  MinNuisance,
  TargetEvent,
  TargetTime,
  Regime,
  PropScoreBackend,
  Censored
)
}
\arguments{
\item{Data}{data.table}

\item{CovDataTable}{data.table}

\item{Model}{list}

\item{CVFolds}{: list}

\item{MinNuisance}{numeric}

\item{TargetEvent}{numeric vector}

\item{TargetTime}{numeric vector}

\item{Regime}{list}

\item{PropScoreBackend}{character}

\item{Censored}{boolean}
}
\description{
Title
}
