% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutput.R
\name{getOutput}
\alias{getOutput}
\title{getOutput}
\usage{
getOutput(Estimate, Estimand = c("RD", "RR", "Risk"))
}
\arguments{
\item{Estimate}{list : a "ConcreteEst" object}

\item{Estimand}{character (default: c("RD", "RR", "Risk")) or a function or a list of functions}
}
\value{
tbd
}
\description{
getOutput
}
\examples{
# getOutput returns risk difference, relative risk, and treatment-specific risks
# concrete.out <- getOutput(Estimate = concrete.est, Estimand = c("rd", "rr", "risk"), 
#                           TargetTime = target.time, TargetEvent = target.event, GComp = TRUE)
# concrete.out$RD
# concrete.out$RR
# concrete.out$Risk
}
