% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutput.R
\name{getOutput}
\alias{getOutput}
\title{getOutput}
\usage{
getOutput(Estimate, Estimand, TargetTime, TargetEvent, GComp)
}
\arguments{
\item{Estimate}{: list concrete Estimate object}

\item{Estimand}{: character or list of function(s)}

\item{TargetTime}{: numeric}

\item{TargetEvent}{: numeric}

\item{GComp}{: boolean}
}
\value{
tbd
}
\description{
getOutput
}
\examples{
"tbd"
}
