% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutput.R
\name{getOutput}
\alias{getOutput}
\title{getOutput}
\usage{
getOutput(Estimate, Estimand = c("RD", "RR", "Risk"), A1 = 1, A0 = 2)
}
\arguments{
\item{Estimate}{list : a "ConcreteEst" object}

\item{Estimand}{character: a subset of c("RD", "RR", "Risk") specifying the target estimand}

\item{A1}{default = 1, the ConcreteEst list element that is the "Treated" arm for risk comparisons}

\item{A0}{default = 2, the ConcreteEst list element that is the "Control" arm for risk comparisons}
}
\value{
data.table of point estimates and standard deviations
}
\description{
getOutput
}
\examples{
# getOutput returns risk difference, relative risk, and treatment-specific risks
# concrete.out <- getOutput(Estimate = concrete.est, Estimand = c("rd", "rr", "risk"),
#                           TargetTime = target.time, TargetEvent = target.event, GComp = TRUE)
# concrete.out$RD
# concrete.out$RR
# concrete.out$Risk

}
