% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutput.R
\name{getOutput}
\alias{getOutput}
\title{getOutput}
\usage{
getOutput(
  Estimate,
  Estimand = c("RD", "RR", "Risk"),
  TargetTime,
  TargetEvent,
  GComp
)
}
\arguments{
\item{Estimate}{: list concrete Estimate object}

\item{Estimand}{: character or list of function(s)}

\item{TargetTime}{: numeric}

\item{TargetEvent}{: numeric}

\item{GComp}{: boolean}
}
\value{
tbd
}
\description{
getOutput
}
\examples{
# getOutput returns risk difference, relative risk, and treatment-specific risks
# concrete.out <- getOutput(Estimate = concrete.est, Estimand = c("rd", "rr", "risk"), 
#                           TargetTime = target.time, TargetEvent = target.event, GComp = TRUE)
# concrete.out$RD
# concrete.out$RR
# concrete.out$Risk
}
