% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPropScore.R
\name{getPropScore}
\alias{getPropScore}
\title{Title}
\usage{
getPropScore(
  Treatment,
  CovDataTable,
  Model,
  MinNuisance,
  Regime,
  PropScoreBackend,
  CVFolds,
  TrtLoss = NULL
)
}
\arguments{
\item{Treatment}{numeric vector}

\item{CovDataTable}{data.table}

\item{Model}{list}

\item{MinNuisance}{numeric}

\item{Regime}{list}

\item{PropScoreBackend}{character}

\item{CVFolds}{list}

\item{TrtLoss}{character or function(A, g.A)}
}
\description{
Title
}
