% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPropScore.R
\name{getPropScore}
\alias{getPropScore}
\title{Title}
\usage{
getPropScore(
  Data,
  CovDataTable,
  Models,
  MinNuisance,
  RegsOfInterest,
  PropScoreBackend,
  CVFolds,
  TrtLoss = NULL
)
}
\arguments{
\item{Data}{data.table}

\item{CovDataTable}{data.table}

\item{Models}{list}

\item{MinNuisance}{numeric}

\item{RegsOfInterest}{list}

\item{PropScoreBackend}{character}

\item{CVFolds}{list}

\item{TrtLoss}{character or function(A, g.A)}
}
\description{
Title
}
