% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doTmleUpdate.R
\name{updateHazard}
\alias{updateHazard}
\title{Update hazards based on clever covariate and PnEIC}
\usage{
updateHazard(
  GStar,
  Hazards,
  TotalSurv,
  NuisanceWeight,
  EvalTimes,
  T.tilde,
  Delta,
  PnEIC,
  NormPnEIC,
  OneStepEps,
  TargetEvent,
  TargetTime
)
}
\arguments{
\item{GStar}{numeric vector of star probabilities}

\item{Hazards}{list of hazard matrices}

\item{TotalSurv}{matrix of survival probabilities}

\item{NuisanceWeight}{matrix of nuisance weights}

\item{EvalTimes}{numeric vector of evaluation times}

\item{T.tilde}{numeric vector of observed event times}

\item{Delta}{numeric vector of observed event types}

\item{PnEIC}{data.table with PnEIC values}

\item{NormPnEIC}{numeric norm of PnEIC}

\item{OneStepEps}{numeric step size for updates}

\item{TargetEvent}{numeric vector of target events}

\item{TargetTime}{numeric vector of target times}
}
\value{
list of updated hazard matrices
}
\description{
Update hazards based on clever covariate and PnEIC
}
